# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: Prognos API
  version: 0.1.0
  description: OpenAPI spec for the Prognos backend (health, readiness, uploads, frameworks, mappings, project summary).
servers:
  - url: http://localhost:4000
    description: Local dev
paths:
...
  /frameworks/{id}/controls:
    get:
      summary: Get controls for a framework
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Framework id
      responses:
        '200':
          description: Controls list
          content:
            application/json:
              schema:
                type: object
                properties:
                  controls:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
        '404':
          description: Framework not found
  /uploads:
    post:
      summary: Create an upload presigned URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
      responses:
        '200':
          description: Upload URL and metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid request
...
  /projects/{id}/summary:
    get:
      summary: Project summary (counts)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Summary object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummary'
components:
  schemas:
...
    Control:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
    UploadRequest:
      type: object
      required:
        - filename
        - projectId
      properties:
        filename: { type: string }
        projectId: { type: string }
        contentType: { type: string }
    UploadResponse:
      type: object
      properties:
        uploadId: { type: string }
        url:
          oneOf:
            - type: string
            - type: "null"
        objectKey: { type: string }
...
    ProjectSummary:
      type: object
      properties:
        projectId: { type: string }
        totalEvidence: { type: integer }
        indexedEvidence: { type: integer }
        mappingsCount: { type: integer }
        controlsTotal: { type: integer }
        controlsCovered: { type: integer }
        timestamp: { type: string, format: date-time }